+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- Adds missing thor dependency for CLI
+- CLI automatically sets base-url for http(s) resources
 
 ## [0.0.10] - 2025-08-01