+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## [Unreleased]
 
+
+## [0.0.11] - 2025-08-05
+
 - Adds missing thor dependency for CLI
 - CLI automatically sets base-url for http(s) resources
 - Added `--http-headers` CLI option for sending custom headers when fetching the initial SOURCE HTML.